requirements.txt
Flask>=2.0
Pillow>=9.0
python-docx>=0.8.11
python-pptx>=0.6.21
reportlab>=3.6.0
Werkzeug>=2.0
